requirements.txt
langchain>=0.2.0,<0.3.0
langchain-community==0.2.0
python-dotenv
streamlit
beautifulsoup4
pypdf
faiss-cpu
groq
fastapi
uvicorn
sse_starlette
googletrans
typer==0.09.0
wikipedia-api
transformers
sentence-transformers
requests
aiocache
elasticsearch
